fix: 修复了定时放卡居然不能设置浮点数的bug

+++ function/core/qmw_editor_of_battle_plan.py
@@ -1123,7 +1123,7 @@ class QMWEditorOfBattlePlan(QMainWindow):
             self.fresh_all_ui()
             return
 
-        ui_value = int(self.EditorAction.WidgetTimeInput.value())
+        ui_value = float(self.EditorAction.WidgetTimeInput.value())
         if e.trigger.time != ui_value:
             e.trigger.time = ui_value
             self.fresh_all_ui()
